Separate random generation from the evaluation
This moves the random generation into its own module, and moves the random
generator script into its own binary instead of replacing the bytebeat main.
